+++ src/main/java/roomescape/admin/AdminController.java
@@ -1,21 +1,21 @@
-package roomescape;
+package roomescape.admin;
 
 import java.net.URI;
 
 import jakarta.validation.Valid;
 
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
 
 import roomescape.reservation.application.ReservationService;
 import roomescape.reservation.dto.ReservationCreateRequest;
 import roomescape.reservation.dto.ReservationResponse;
 
-@Controller
+@RestController
+@RequestMapping("/admin")
 public class AdminController {
 
     private final ReservationService reservationService;
@@ -24,30 +24,9 @@ public class AdminController {
         this.reservationService = reservationService;
     }
 
-    @GetMapping("/")
-    public String index() {
-        return "admin/index";
-    }
-
-    @GetMapping("/admin/reservation")
-    public String reservation() {
-        return "admin/reservation";
-    }
-
-    @PostMapping("/admin/reservations")
-    @ResponseBody
+    @PostMapping("/reservations")
     public ResponseEntity<ReservationResponse> reservationPost(@RequestBody @Valid ReservationCreateRequest request) {
         ReservationResponse response = reservationService.save(request, request.memberId());
         return ResponseEntity.created(URI.create("/reservations/" + response.id())).body(response);
-    }
-
-    @GetMapping("/admin/time")
-    public String time() {
-        return "admin/time";
-    }
-
-    @GetMapping("/admin/theme")
-    public String theme() {
-        return "admin/theme";
     }
 }
