+++ src/main/java/roomescape/page/PageController.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.page;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
